+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Add `http.request.method` attribute for performance span data ([#786](https://github.com/getsentry/sentry-go/pull/786))
 - Automatic transactions for Fasthttp integration ([#732](https://github.com/getsentry/sentry-go/pull/723))
 - Add `Fiber` integration ([#795](https://github.com/getsentry/sentry-go/pull/795))
+- Use `errors.Unwrap()` to create exception groups ([#792](https://github.com/getsentry/sentry-go/pull/792))
 
 ## 0.27.0